Progress commit
Added relaxation method


Added relaxation method

+++ Lab1/src/main/kotlin/Method.kt
@@ -1,3 +1,3 @@
 interface Method {
-    fun run(): Float
+    fun run(): Double
 }
